Flatten the namespace hierarchy for commands, queries, API endpoints, and tests
