--- account-management/Api/Tenants/Contracts/CreateTenantRequest.cs
@@ -1,6 +1,0 @@
-using JetBrains.Annotations;
-
-namespace PlatformPlatform.AccountManagement.Api.Tenants.Contracts;
-
-[UsedImplicitly]
-public sealed record CreateTenantRequest(string Name, string Subdomain, string Email, string? Phone);
