--- account-management/Api/Tenants/Contracts/UpdateTenantRequest.cs
@@ -1,3 +1,0 @@
-namespace PlatformPlatform.AccountManagement.Api.Tenants.Contracts;
-
-public sealed record UpdateTenantRequest(string Name, string Email, string? Phone);
