+++ account-management/Api/Tenants/TenantContractsV1.cs
@@ -1,7 +1,12 @@
 using JetBrains.Annotations;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
 
-namespace PlatformPlatform.AccountManagement.Api.Tenants.Contracts;
+namespace PlatformPlatform.AccountManagement.Api.Tenants;
+
+[UsedImplicitly]
+public sealed record CreateTenantRequest(string Name, string Subdomain, string Email, string? Phone);
+
+public sealed record UpdateTenantRequest(string Name, string Email, string? Phone);
 
 [UsedImplicitly(ImplicitUseTargetFlags.Members)]
 public sealed record TenantResponseDto
