--- account-management/Api/Users/Contracts/CreateUserRequest.cs
@@ -1,7 +1,0 @@
-using JetBrains.Annotations;
-using PlatformPlatform.AccountManagement.Domain.Users;
-
-namespace PlatformPlatform.AccountManagement.Api.Users.Contracts;
-
-[UsedImplicitly]
-public sealed record CreateUserRequest(string Email, UserRole UserRole);
