--- account-management/Api/Users/Contracts/UpdateUserRequest.cs
@@ -1,5 +1,0 @@
-using PlatformPlatform.AccountManagement.Domain.Users;
-
-namespace PlatformPlatform.AccountManagement.Api.Users.Contracts;
-
-public sealed record UpdateUserRequest(string Email, UserRole UserRole);
