+++ account-management/Api/Users/UserContractsV1.cs
@@ -1,7 +1,12 @@
 using JetBrains.Annotations;
 using PlatformPlatform.AccountManagement.Domain.Users;
 
-namespace PlatformPlatform.AccountManagement.Api.Users.Contracts;
+namespace PlatformPlatform.AccountManagement.Api.Users;
+
+[UsedImplicitly]
+public sealed record CreateUserRequest(string Email, UserRole UserRole);
+
+public sealed record UpdateUserRequest(string Email, UserRole UserRole);
 
 [UsedImplicitly(ImplicitUseTargetFlags.Members)]
 public sealed record UserResponseDto
