+++ account-management/Application/Tenants/GetTenant.cs
@@ -2,7 +2,7 @@ using MediatR;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
 using PlatformPlatform.SharedKernel.ApplicationCore.Cqrs;
 
-namespace PlatformPlatform.AccountManagement.Application.Tenants.Queries;
+namespace PlatformPlatform.AccountManagement.Application.Tenants;
 
 public static class GetTenant
 {
