+++ account-management/Application/Tenants/TenantCreatedEventHandler.cs
@@ -3,7 +3,7 @@ using MediatR;
 using Microsoft.Extensions.Logging;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
 
-namespace PlatformPlatform.AccountManagement.Application.Tenants.EventHandlers;
+namespace PlatformPlatform.AccountManagement.Application.Tenants;
 
 [UsedImplicitly]
 public sealed class TenantCreatedEventHandler : INotificationHandler<TenantCreatedEvent>
