+++ account-management/Application/Tenants/TenantValidator.cs
@@ -2,7 +2,7 @@ using FluentValidation;
 using JetBrains.Annotations;
 using PlatformPlatform.SharedKernel.ApplicationCore.Validation;
 
-namespace PlatformPlatform.AccountManagement.Application.Tenants.Commands;
+namespace PlatformPlatform.AccountManagement.Application.Tenants;
 
 public interface ITenantValidation
 {
