+++ account-management/Application/Tenants/UpdateTenant.cs
@@ -3,7 +3,7 @@ using MediatR;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
 using PlatformPlatform.SharedKernel.ApplicationCore.Cqrs;
 
-namespace PlatformPlatform.AccountManagement.Application.Tenants.Commands;
+namespace PlatformPlatform.AccountManagement.Application.Tenants;
 
 public static class UpdateTenant
 {
