+++ account-management/Application/Users/GetUser.cs
@@ -3,7 +3,7 @@ using MediatR;
 using PlatformPlatform.AccountManagement.Domain.Users;
 using PlatformPlatform.SharedKernel.ApplicationCore.Cqrs;
 
-namespace PlatformPlatform.AccountManagement.Application.Users.Queries;
+namespace PlatformPlatform.AccountManagement.Application.Users;
 
 public static class GetUser
 {
