+++ account-management/Application/Users/UpdateUser.cs
@@ -3,7 +3,7 @@ using MediatR;
 using PlatformPlatform.AccountManagement.Domain.Users;
 using PlatformPlatform.SharedKernel.ApplicationCore.Cqrs;
 
-namespace PlatformPlatform.AccountManagement.Application.Users.Commands;
+namespace PlatformPlatform.AccountManagement.Application.Users;
 
 public static class UpdateUser
 {
