+++ account-management/Tests/Application/Tenants/TenantCreatedEventHandlerTests.cs
@@ -6,12 +6,11 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 using PlatformPlatform.AccountManagement.Application;
-using PlatformPlatform.AccountManagement.Application.Tenants.Commands;
-using PlatformPlatform.AccountManagement.Application.Tenants.EventHandlers;
+using PlatformPlatform.AccountManagement.Application.Tenants;
 using PlatformPlatform.AccountManagement.Infrastructure;
 using Xunit;
 
-namespace PlatformPlatform.AccountManagement.Tests.Application.Tenants.EventHandlers;
+namespace PlatformPlatform.AccountManagement.Tests.Application.Tenants;
 
 public sealed class TenantCreatedEventHandlerTests : IDisposable
 {
