+++ shared-kernel/Tests/InfrastructureCore/Persistence/RepositoryTests.cs
@@ -133,7 +133,8 @@ public sealed class RepositoryTests : IDisposable
         primaryRepository.Update(originalTestAggregate);
 
         // Assert
-        await Assert.ThrowsAsync<DbUpdateConcurrencyException>(() => _testDbContext.SaveChangesAsync(cancellationToken));
+        await Assert.ThrowsAsync<DbUpdateConcurrencyException>(() =>
+            _testDbContext.SaveChangesAsync(cancellationToken));
     }
 
     [Fact]
